% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_nuts_codes.R
\docType{data}
\name{all_nuts_codes}
\alias{all_nuts_codes}
\title{List of all NUTS codes and classifications}
\format{
\subsection{\code{all_nuts_codes}}{

A data frame with 8,896 rows and 2 columns:
\describe{
\item{code}{NUTS code}
\item{version}{NUTS versions}
\item{country}{Country name}
}
}
}
\source{
\url{https://urban.jrc.ec.europa.eu/tools/nuts-converter?lng=en#/}
}
\usage{
all_nuts_codes
}
\description{
The data frame stores all NUTS codes in hierarchical levels 1, 2 and 3 by
NUTS classification versions 2006, 2010, 2013, 2016 and 2021.
}
\keyword{datasets}
