% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutsDual.R
\name{BuildTree}
\alias{BuildTree}
\title{Build trees}
\usage{
BuildTree(theta, r, u, v, j, epsilon, L, joint0)
}
\arguments{
\item{theta}{starting position}

\item{r}{starting momentum}

\item{u}{a slice variable, which is used to simplify the implementation}

\item{v}{a randomly generated value, which is used to determine the direction of movement}

\item{j}{the height of a tree}

\item{epsilon}{step size}

\item{joint0}{to avoid computing the same value repeatedly, we compute joint0=log0- 0.5*(r0 %*% r0) beforehand and take this value as an input for the BUildTree function}
}
\value{
a list of variables. It includes updated leftmost and rightmost states. It also returns updated values which are used to improve memory efficience.
}
\description{
This function builds the tree for NUTS
}
