% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_nuts.R
\name{classify_nuts}
\alias{classify_nuts}
\title{Classify version of NUTS codes}
\usage{
classify_nuts(
  data = data,
  nuts_code = nuts_code,
  group_vars = NULL,
  ties = "most_recent",
  quiet = FALSE
)
}
\arguments{
\item{data}{A data frame or tibble that contains a variable with NUTS \code{1}, \code{2} or \code{3} codes and possibly other variables.
NUTS codes must be of the same level and need to be unique, unless additional grouping variables are specified. No
duplicate NUTS codes within groups allowed.}

\item{nuts_code}{Variable name containing NUTS codes}

\item{group_vars}{Variable name(s) for classification within groups. Always computes overlap within country. \code{NULL}by default.}

\item{ties}{Picks \code{'most_recent'} or \code{'oldest'} version when overlap is identical across multiple NUTS versions. \code{'most_recent'}
by default.}

\item{quiet}{Suppress messages and warnings. \code{TRUE} by default.}
}
\value{
A list of three tibbles. The first tibble contains the original data with the classified NUTS version, level, and country.
The second tibble lists the group-specific overlap with each NUTS version. The third tibble shows missing NUTS codes
for each group.

The output can be passed to \code{\link[=convert_nuts_version]{convert_nuts_version()}} to convert data across NUTS versions and \code{\link[=convert_nuts_level]{convert_nuts_level()}} to aggregate across NUTS levels.
}
\description{
\code{classify_nuts()} can identify the NUTS version year and level from a variable containing NUTS codes
}
\examples{
library(dplyr)

# Load EUROSTAT data of manure storage deposits
data(manure)

# Data varies at the NUTS level x indicator x year x country x NUTS code level
head(manure)

# Classify version of NUTS 2 codes in Germany
manure \%>\%
 filter(nchar(geo) == 4) \%>\%
 filter(indic_ag == 'I07A_EQ_Y') \%>\%
 filter(grepl('^DE', geo)) \%>\%
 filter(time == 2003) \%>\%
 select(-indic_ag, -time) \%>\%
 # Data varies at the NUTS code level
 classify_nuts(nuts_code = 'geo')

# Classify version of NUTS 3 codes within country and year
manure \%>\%
  filter(nchar(geo) == 5) \%>\%
  filter(indic_ag == 'I07A_EQ_Y') \%>\%
  select(-indic_ag) \%>\%
  # Data varies at the year x country x NUTS code level. The country grouping
  # is always used by default.
  classify_nuts(nuts_code = 'geo', group_vars = 'time')


}
