% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmcDual.R, R/NutsDual.R
\name{FindReasonableEpsilon}
\alias{FindReasonableEpsilon}
\alias{FindReasonableEpsilon}
\title{FindReasonableEpsilon}
\usage{
FindReasonableEpsilon(theta0, log.start, L)

FindReasonableEpsilon(theta0, log.start, L)
}
\arguments{
\item{log.start}{the log posterior value at initial state}

\item{L}{callable function needed in Leapfrog}

\item{theta}{initial state}

\item{grad.start}{the gradient value at initial state}

\item{theta}{initial state}

\item{log.start}{the log posterior value at initial state}

\item{grad.start}{the gradient value at initial state}

\item{L}{callable function needed in Leapfrog}
}
\value{
initial epsilon

initial epsilon
}
\description{
Heuristic for choosing an initial value of epsilon

Heuristic for choosing an initial value of epsilon
}
