% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTStest.R
\name{g}
\alias{g}
\title{g}
\usage{
g(...)
}
\description{
The function is copied from "https://strugglingthroughproblems.wordpress.com/2010/08/27/matlab-style-multiple-assignment-in%C2%A0r/".
This function grabs '...' part without evaluating it (so pass by name)
}
