% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutsDual.R, R/hmc.R
\name{\%=\%.lbunch}
\alias{\%=\%.lbunch}
\alias{\%=\%.lbunch}
\title{%=%,lbunch}
\usage{
\method{\%=\%}{lbunch}(l, r, ...)

\method{\%=\%}{lbunch}(l, r, ...)
}
\arguments{
\item{l}{the list on the left-hand side}

\item{r}{the list on the right-hand side}

\item{l}{the list on the left-hand side}

\item{r}{the list on the right-hand side}
}
\description{
The function is copied from "https://strugglingthroughproblems.wordpress.com/2010/08/27/matlab-style-multiple-assignment-in%C2%A0r/"
It is a specific implementation for the generic %=%

The function is copied from "https://strugglingthroughproblems.wordpress.com/2010/08/27/matlab-style-multiple-assignment-in%C2%A0r/"
It is a specific implementation for the generic %=%
}
