% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutsDual.R
\name{\%=\%}
\alias{\%=\%}
\title{%=%}
\usage{
l \%=\% r
}
\arguments{
\item{l}{the list on the left-hand side}

\item{r}{the list on the right-hand side}
}
\description{
The function is copied from "https://strugglingthroughproblems.wordpress.com/2010/08/27/matlab-style-multiple-assignment-in%C2%A0r/" and is used for multiple assignment
}
