% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{hmc.nograd}
\alias{hmc.nograd}
\title{Hamiltonian Monte Carlo}
\usage{
hmc.nograd(theta.start, epsilon, L, logDensity, dlogDensity, M)
}
\arguments{
\item{theta.start}{a p-dimensional vector with the initial value for each parameter.}

\item{epsilon}{a non-negative value specifying the leapfrog step size.}

\item{L}{an integer number specifying the length of the leapfrog step.}

\item{logDensity}{a function that computes the logarithm of the joint density of the variables.}

\item{M}{an integer specifying the number of iterations.}
}
\value{
This function returns a matrix which m-th row is a sample from the joint density.
}
\description{
Hamiltonian Monte Carlo
}
