% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmc.R
\name{Hmc}
\alias{Hmc}
\title{Hmc}
\usage{
Hmc(theta0, epsilon, leap.nsteps, L, M)
}
\arguments{
\item{theta0}{a p-dimensional vector with the initial value for each parameter.}

\item{epsilon}{a non-negative value specifying the leapfrog step size.}

\item{leap.nsteps}{an integer number specifying the number of leapfrog steps.}

\item{L}{callable function: returns the value of log posterior and the gradient of log posterior probability at given input.}

\item{M}{an integer specifying the number of iterations.}
}
\value{
This function returns a matrix whose m-th row is a sample from the joint density.
}
\description{
This function performs the Hamiltonian Monte Carlo.
}
