% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HmcDual.R
\name{HmcDual}
\alias{HmcDual}
\title{HmcDual}
\usage{
HmcDual(theta0, delta, lambda, L, M, Madapt)
}
\arguments{
\item{theta0}{a p-dimensional vector with the initial value for each parameter.}

\item{delta}{the desired average acceptance probability}

\item{lambda}{trajectory length for leapfrog}

\item{L}{a callable function that returns a 2-dimensional vector whose first element is the
logarithm of the density of the input and the second is the gradient of the logarithm
of the density evaluated at the input.}

\item{M}{an integer specifying the number of iterations.}

\item{Madapt}{an integer specifying the number of iterations of the warmup phase.}
}
\value{
This function returns a matrix whose m-th row is a sample from the joint density.
}
\description{
Hamiltonian Monte Carlo with Dual Averaging
}
