% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{Leapfrog}
\alias{Leapfrog}
\alias{Leapfrog}
\title{Leapfrog}
\usage{
Leapfrog(theta, r, epsilon, L)

Leapfrog(theta, r, epsilon, L)
}
\arguments{
\item{theta}{starting position}

\item{r}{starting momentum}

\item{epsilon}{step size}

\item{L}{callable function: returns the value of log posterior and the gradient of log posterior prbability at given input}

\item{theta}{starting position}

\item{r}{starting momentum}

\item{epsilon}{step size}

\item{L}{callable function: returns the value of log posterior and the gradient of log posterior prbability at given input}
}
\description{
Leapfrog

Leapfrog
}
