% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.r
\docType{package}
\name{NUTS-package}
\alias{NUTS-package}
\title{No-U-Turn sampler for Hamiltonian Monte Carlo}
\description{
A pure R implementation of the No-U-Turn sampler for Hamiltonian
Monte Carlo (algorithm 6 of Hoffman and Gelman 2014).
}
\references{
Hoffman, M. D., & Gelman, A. (2014). The No-U-turn sampler:
  adaptively setting path lengths in Hamiltonian Monte
  Carlo. Journal of Machine Learning Research, 15(1), 1593-1623.
}
\author{
Kaspar Märtens, S. Wotherspoon
}
