% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{NutsDual}
\alias{NutsDual}
\alias{NutsDual}
\title{NutsDual}
\usage{
NutsDual(theta0, delta, L, M, Madapt)

NutsDual(theta0, delta, L, M, Madapt)
}
\arguments{
\item{theta0}{inital state of theta}

\item{delta}{the desired average acceptance probability}

\item{L}{a callable function that returns log probability}

\item{M}{number of samples to generate}

\item{Madapt}{number of iterations

This function implements No-U-Turn Sampler with Dual Averaging.
The code is a modified version of algorithm6 in paper. Instead of including samples in burn-in period (m < Madapt), the function generates a sample where samples in the burn-in period are discarded.}

\item{theta0}{inital state of theta}

\item{delta}{the desired average acceptance probability}

\item{L}{a callable function that returns log probability}

\item{M}{number of samples to generate}

<<<<<<< HEAD
\item{Madapt}{number of iterations}
=======
\item{Madapt}{number of iterations

This function implements No-U-Turn Sampler with Dual Averaging.
The code is a modified version of algorithm6 in paper. Instead of including samples in burn-in period (m < Madapt), the function generates a sample where samples in the burn-in period are discarded.}
>>>>>>> 6ef3cd0bb9ad536eb744d4070deafb4a3c451200
}
\value{
samples

<<<<<<< HEAD
samples generated by NUTS
=======
samples
>>>>>>> 6ef3cd0bb9ad536eb744d4070deafb4a3c451200
}
\description{
NutsDual

<<<<<<< HEAD
This function implements No-U-Turn Sampler with Dual Averaging.
The code is a modified version of algorithm6 in paper. Instead of including samples in burn-in period (m < Madapt), the function generates a sample where samples in the burn-in period are discarded.
=======
NutsDual
>>>>>>> 6ef3cd0bb9ad536eb744d4070deafb4a3c451200
}
