% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stitch.R
\name{stitch}
\alias{stitch}
\title{Stitch recent screenshots together into a gif}
\usage{
stitch(n, outfile, outdir = getwd(),
  dir = "/Users/anthony/Dropbox/Screenshots", fps = 1)
}
\arguments{
\item{n}{number of recent screenshots}

\item{outfile}{name of output file}

\item{dir}{screenshot directory}

\item{fps}{frames per second for resulting gif}
}
\value{
full path to gif (also in your clipboard)
}
\description{
Stitch recent screenshots together into a gif
}
