% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{BuildTree}
\alias{BuildTree}
\alias{BuildTree}
\title{Build trees}
\usage{
BuildTree(theta, r, u, v, j, epsilon, L, joint0)

BuildTree(theta, r, u, v, j, epsilon, L, joint0)
}
\arguments{
\item{j}{the height of a tree}

\item{epsilon}{step size}

\item{joint0}{to avoid computing the same value repeatedly, we compute joint0=log0- 0.5*(r0 %*% r0) beforehand and take this value as an input for the BUildTree function

This function builds the tree for NUTS}

\item{j}{the height of a tree}

\item{epsilon}{step size}

\item{joint0}{to avoid computing the same value repeatedly, we compute joint0=log0- 0.5*(r0 %*% r0) beforehand and take this value as an input for the BUildTree function

This function builds the tree for NUTS}
}
\description{
Build trees

Build trees
}
