% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NutsDual.R
\name{FindReasonableEpsilon}
\alias{FindReasonableEpsilon}
\title{FindReasonableEpsilon}
\usage{
FindReasonableEpsilon(theta, log.start, L)
}
\arguments{
\item{log.start}{the log posterior value at initial state}

\item{L}{callable function needed in Leapfrog}

\item{grad.start}{the gradient value at initial state}
}
\value{
initial epsilon
}
\description{
Heuristic for choosing an initial value of epsilon
}
