% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{FindReasonableEpsilon}
\alias{FindReasonableEpsilon}
\alias{FindReasonableEpsilon}
\title{FindReasonableEpsilon}
\usage{
FindReasonableEpsilon(theta, log.start, L)

FindReasonableEpsilon(theta, log.start, L)
}
\arguments{
\item{log.start}{the log posterior value at initial state}

\item{L}{callable function needed in Leapfrog}

\item{grad.start}{the gradient value at initial state}

\item{log.start}{the log posterior value at initial state}

\item{grad.start}{the gradient value at initial state}

\item{L}{callable function needed in Leapfrog}
}
\value{
epsilon

Heuristic for choosing an initial value of epsilon

initial epsilon
}
\description{
FindReasonableEpsilon

Heuristic for choosing an initial value of epsilon
}
