% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts-package.R
\docType{package}
\name{nuts-package}
\alias{nuts}
\alias{nuts-package}
\title{nuts: Convert European Regional Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Motivated by changing administrative boundaries over time, the 'nuts' package can convert European regional data with NUTS codes between versions (2006, 2010, 2013, 2016 and 2021) and levels (NUTS 1, NUTS 2 and NUTS 3). The package uses spatial interpolation as in Lam (1983) \doi{10.1559/152304083783914958} based on granular (100m x 100m) area, population and land use data provided by the European Commission's Joint Research Center.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/nuts/}
  \item \url{https://github.com/ropensci/nuts}
  \item Report bugs at \url{https://github.com/ropensci/nuts}
}

}
\author{
\strong{Maintainer}: Moritz Hennicke \email{AAoritz@posteo.de} (\href{https://orcid.org/0000-0001-6811-1821}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Werner Krause \email{werner.krause@uni-potsdam.de} (\href{https://orcid.org/0000-0002-5069-7964}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Pueyo-Ros Josep (Josep reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/623#issuecomment-1951446662) [reviewer]
  \item Le Meur Nolwenn (Nolwenn reviewed the package for rOpenSci, see https://github.com/ropensci/software-review/issues/623#issuecomment-1961501137) [reviewer]
}

}
\keyword{internal}
