% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{NUTS}
\alias{NUTS}
\title{No-U-Turn sampler}
\usage{
NUTS(
  theta,
  f,
  grad_f,
  n_iter,
  M = diag(length(theta)),
  M_adapt = 50,
  M_chol = Matrix::chol(M),
  M_inv = Matrix::solve(M),
  delta = 0.5,
  max_treedepth = 10,
  eps = 1,
  find = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{theta}{Initial value for the parameters}

\item{f}{log-likelihood function (up to a constant)}

\item{grad_f}{the gradient of the log-likelihood function}

\item{n_iter}{Number of MCMC iterations}

\item{M}{the HMC mass matrix. Defaults to \code{diag(length(theta))}.}

\item{M_adapt}{Parameter M_adapt in algorithm 6 in the NUTS paper}

\item{M_chol}{The cholesky decomposition of \code{M}. Default \code{Matrix::chol(M)}.}

\item{M_inv}{The inverse of \code{M}. Default \code{Matrix::solve(M)}.}

\item{delta}{Target acceptance ratio, defaults to 0.5}

\item{max_treedepth}{Maximum depth of the binary trees constructed by NUTS}

\item{eps}{Starting guess for epsilon}

\item{find}{Allow NUTS to make a better guess at epsilon? Default \code{TRUE}.}

\item{verbose}{logical. Message diagnostic info each iteration? Default \code{TRUE}.}
}
\value{
Matrix with the trace of sampled parameters. Each mcmc iteration in rows and parameters in columns.
}
\description{
No-U-Turn sampler
}
