% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts.R
\name{NUTS}
\alias{NUTS}
\title{No-U-Turn sampler}
\usage{
NUTS(theta, f, grad_f, n_iter, M_diag = NULL, M_adapt = 50, delta = 0.5,
  max_treedepth = 10, eps = 1, verbose = TRUE)
}
\arguments{
\item{theta}{Initial value for the parameters}

\item{f}{log-likelihood function (up to a constant)}

\item{grad_f}{the gradient of the log-likelihood function}

\item{n_iter}{Number of MCMC iterations}

\item{M_diag}{Diagonal elements of the mass matrix in HMC. Defaults to ones.}

\item{M_adapt}{Parameter M_adapt in algorithm 6 in the NUTS paper}

\item{delta}{Target acceptance ratio, defaults to 0.5}

\item{max_treedepth}{Maximum depth of the binary trees constructed by NUTS}

\item{eps}{Starting guess for epsilon}
}
\value{
Matrix with the trace of sampled parameters. Each mcmc iteration in rows and parameters in columns.
}
\description{
No-U-Turn sampler
}
