% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{NutsDual}
\alias{NutsDual}
\alias{NutsDual}
\title{NutsDual}
\usage{
NutsDual(theta0, delta, L, M, Madapt)

NutsDual(theta0, delta, L, M, Madapt)
}
\arguments{
\item{theta0}{inital state of theta}

\item{delta}{the desired average acceptance probability}

\item{L}{a callable function that returns log probability}

\item{M}{number of samples to generate}

\item{Madapt}{number of iterations

This function implements No-U-Turn Sampler with Dual Averaging.
The code is a modified version of algorithm6 in paper. Instead of including samples in burn-in period (m < Madapt), the function generates a sample where samples in the burn-in period are discarded.}

\item{theta0}{inital state of theta}

\item{delta}{the desired average acceptance probability}

\item{L}{a callable function that returns log probability}

\item{M}{number of samples to generate}

\item{Madapt}{number of iterations}
}
\value{
samples

samples generated by NUTS
}
\description{
NutsDual

This function implements No-U-Turn Sampler with Dual Averaging.
The code is a modified version of algorithm6 in paper. Instead of including samples in burn-in period (m < Madapt), the function generates a sample where samples in the burn-in period are discarded.
}
