% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NUTS.R, R/NUTStest.R
\name{StopCon}
\alias{StopCon}
\alias{StopCon}
\title{StopCon}
\usage{
StopCon(theta.minus, theta.plus, r.minus, r.plus)

StopCon(theta.minus, theta.plus, r.minus, r.plus)
}
\arguments{
\item{r.plus}{This function computes the U-Turn stopping condition, which is used in NUTS and BuildTree functions}

\item{r.plus}{This function computes the U-Turn stopping condition, which is used in NUTS and BuildTree functions}
}
\description{
StopCon

StopCon
}
